requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
boto3=1.24.8=pyhd8ed1ab_0
botocore=1.27.8=pyhd8ed1ab_0
brotlipy=0.7.0=py310h1961e1f_1004
bzip2=1.0.8=h0d85af4_4
ca-certificates=2022.5.18.1=h033912b_0
certifi=2022.5.18.1=py310h2ec42d9_0
cffi=1.15.0=py310hcc37b68_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.1.3=py310h2ec42d9_0
cryptography=37.0.1=py310hf6deb26_0
idna=3.3=pyhd8ed1ab_0
jmespath=1.0.0=pyhd8ed1ab_0
libffi=3.4.2=h0d85af4_5
libzlib=1.2.12=h6c3fc93_0
ncurses=6.3=h96cf925_1
openssl=3.0.3=hfe4f2af_0
pip=22.1.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pysocks=1.7.1=py310h2ec42d9_5
python=3.10.5=hdd68b96_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.20.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2022.1=pyhd8ed1ab_0
readline=8.1.2=h3899abd_0
requests=2.28.0=pyhd8ed1ab_0
s3transfer=0.6.0=pyhd8ed1ab_0
setuptools=62.3.4=py310h2ec42d9_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.38.5=hd9f0692_0
tk=8.6.12=h5dbffcc_0
twilio=7.9.2=pyhd8ed1ab_0
tzdata=2022a=h191b570_0
urllib3=1.26.9=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.5=haf1e3a3_1
zlib=1.2.12=h6c3fc93_0
